% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Extract model entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{A \code{lambdas} object returned by \code{\link{parse_lambdas}}.}
}
\description{
Extract the entropy of a Maxent model from a \code{lambdas} objects generated
by \code{\link{parse_lambdas}}.
}
\examples{
# Below we use the dismo::maxent example to fit a Maxent model:
if (require(dismo) && require(rJava) && 
    file.exists(system.file('java/maxent.jar', package='dismo'))) {
  fnames <- list.files(system.file('ex', package="dismo"), '\\\\.grd$', 
                       full.names=TRUE )
  predictors <- stack(fnames)
  occurrence <- system.file('ex/bradypus.csv', package='dismo')
  occ <- read.table(occurrence, header=TRUE, sep=',')[,-1]
  me <- maxent(predictors, occ, path=file.path(tempdir(), 'example'), 
               factors='biome')

  # ... and then parse the lambdas information:
  lam <- parse_lambdas(me)
  entropy(lam)
}

}
\seealso{
\code{\link{parse_lambdas}} \code{\link{project}}
}
\keyword{lambdas}
\keyword{maxent,}
