% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maxent.R
\name{get_maxent}
\alias{get_maxent}
\title{Download maxent.jar into dismo/java}
\usage{
get_maxent(version = "latest", quiet = FALSE)
}
\arguments{
\item{version}{Either 'latest' for the latest stable release, or a character
version string indicating the version to download. See
\code{maxent_versions} for available versions. Default is \code{'latest'}.}

\item{quiet}{logical. Should system and success messages be suppressed?}
}
\value{
Invisible returns \code{0} for success, and a non-zero integer
otherwise.
}
\description{
\code{get_maxent} downloads maxent.jar.
}
\details{
This downloads the java file for the requested Maxent version into the
appropriate dismo/java folder.

See references for license and citation details.
}
\examples{
\dontrun{
get_maxent('3.3.3k')
}
}
\references{
\itemize{
\item GitHub repository: \href{https://github.com/mrmaxent/Maxent}{mrmaxent/Maxent}
\item Steven J. Phillips, Miroslav Dudík, Robert E. Schapire. \href{https://biodiversityinformatics.amnh.org/open_source/maxent/}{Maxent software for modeling species niches and distributions.}
}
}
\seealso{
\code{\link[=maxent_versions]{maxent_versions()}}
}
