% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Calculate information criteria for Maxent models}
\usage{
ic(x, occ, lambdas)
}
\arguments{
\item{x}{Either a file path to a Maxent raw prediction raster, or a
\code{Raster} or \code{RasterStack} containing raw Maxent predictions for
which information criteria will be calculated.}

\item{occ}{A \code{matrix} or \code{data.frame} containing the coordinates
for occurrence data used to fit the model.}

\item{lambdas}{The path to a Maxent .lambdas file, or a \code{MaxEnt} fitted
model object.}
}
\value{
An n x 6 matrix, where n is the number of Maxent rasters for which
  information criteria are to be calculated. Columns give \code{n} (the
  number of occurrence records used for model training), \code{k} (the number
  of features with non-zero weights), \code{ll} (the negative log likelihood
  of the model), \code{AIC}, \code{AICc}, and \code{BIC} (as calculated in
  ENMTools).
}
\description{
Calculate AIC, AICc, and BIC for Maxent models as implemented in ENMTools.
}
\section{Warning}{
 
These information criteria should \emph{not} be calculated for models that
use hinge or threshold features because the number of predictors is not
estimated correctly.
}
\examples{
library(dismo)
fnames <- list.files(path=file.path(system.file(package='dismo'), 'ex'), 
                     patt='grd', full=TRUE )
predictors <- stack(fnames)
occ <- read.csv(file.path(system.file(package="dismo"), 'ex/bradypus.csv'))[, -1]
me <- maxent(predictors, occ, args=c('hinge=false', 'threshold=false'),
             path=tempdir())
r <- project_maxent(me, predictors, quiet=TRUE)$prediction_raw

# passing the raster object to pred.raw and the maxent object to lambdas:
maxent_IC(r, occ, me)

# passing a lambdas file path to lambdas:
maxent_IC(r, occ, file.path(tempdir(), 'species.lambdas'))

# passing a raster file path and lambdas file path to lambdas:
writeRaster(r, f <- tempfile(fileext='.tif'))
maxent_IC(f, occ, file.path(tempdir(), 'species.lambdas'))
}
\references{
\itemize{
  \item{Warren, D. L., Glor, R. E. and Turelli, M. 2009. \href{http://onlinelibrary.wiley.com/doi/10.1111/j.1600-0587.2009.06142.x/full}{ENMTools: a toolbox for comparative studies of environmental niche models.} \emph{Ecography} 33:607-611}
  \item{\href{http://enmtools.blogspot.com.au/}{ENMTools}}
}
}

