% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_maxent.R
\name{import_maxent}
\alias{import_maxent}
\title{Import a Maxent model into a MaxEnt object}
\usage{
import_maxent(dir, prefix = "species")
}
\arguments{
\item{dir}{The file path to the directory containing Maxent output.}

\item{prefix}{The species name, which forms the prefix of the .lambdas file. 
The default is \code{'species'} (and so species.lambdas is expected to
exist in \code{dir}).}
}
\value{
A \code{MaxEnt} object (see \code{\link{MaxEnt-class}}).
}
\description{
Create a MaxEnt object from a previously-fitted Maxent model
}
\examples{
library(dismo)
predictors <- stack(list.files(
  file.path(system.file(package='dismo'), 'ex'), '\\\\.grd$', full=TRUE))
occ <- read.csv(file.path(system.file(package="dismo"), 'ex/bradypus.csv'))[, -1]
d <- file.path(tempdir(), 'demo')
dir.create(d)
maxent(predictors, occ, factors='biome', path=d)
m <- import_maxent(d)
}

