% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_versions.R
\name{maxent_versions}
\alias{maxent_versions}
\title{Show available versions of Maxent}
\usage{
maxent_versions(include_beta = FALSE)
}
\arguments{
\item{include_beta}{logical. Should beta versions be included?}
}
\value{
Returns a \code{character} vector of version numbers.
}
\description{
\code{maxent_versions} returns a vector of available version numbers.
}
\details{
This obtains a vector of versions numbers for available versions of Maxent,
from https://github.com/mrmaxent/Maxent/tree/master/ArchivedReleases.
}
\examples{
\dontrun{
maxent_versions()
}
}
\seealso{
\code{\link[=get_maxent]{get_maxent()}}
}
