% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_features.R
\name{n_features}
\alias{n_features}
\title{Extract the number of features used in a Maxent model}
\usage{
n_features(lambdas)
}
\arguments{
\item{lambdas}{A \code{vector} of one or more file paths to Maxent .lambdas 
files, or a \code{list} of one or more \code{MaxEnt} fitted model objects.}
}
\value{
An vector with one element for each element of \code{lambdas}.
}
\description{
Calculate the number of non-zero features used in a Maxent model.
}
\examples{
# Below we use the dismo::maxent example to fit a Maxent model:
if (require(dismo) && require(rJava) && 
    file.exists(system.file('java/maxent.jar', package='dismo'))) {
  fnames <- list.files(system.file('ex', package='dismo'), '\\\\.grd$', 
                       full.names=TRUE )
  predictors <- stack(fnames)
  occurrence <- system.file('ex/bradypus.csv', package='dismo')
  occ <- read.table(occurrence, header=TRUE, sep=',')[,-1]
  me <- maxent(predictors, occ, args=c('hinge=false', 'threshold=false'),
               path=file.path(tempdir(), 'ex1'))
  n_features(me)
  
  # comparing multiple models
  me2 <- maxent(predictors, occ, args=c('betamultiplier=3'), 
                path=file.path(tempdir(), 'ex2'))
  n_features(list(me, me2))
  
  # or use lambdas files
  ff <- file.path(tempdir(), c('ex1', 'ex2'), 'species.lambdas')
  ff
  n_features(ff)
}
}
