% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lambdas.R
\name{print.lambdas}
\alias{print.lambdas}
\title{Print rmaxent lambdas objects}
\usage{
\method{print}{lambdas}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{lambdas} object returned by \code{\link[=parse_lambdas]{parse_lambdas()}}.}

\item{digits}{The minimum number of significant digits to be used.}

\item{...}{Other arguments to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\description{
Print method for \code{lambdas} objects generated by \code{\link[=parse_lambdas]{parse_lambdas()}}.
}
\details{
This function summarises the \code{data.frame} holding feature
weights in an \code{rmaxent} \code{lambdas} object. To see that information
in full for \code{lambdas} object \code{x}, use \code{x$lambdas} or
\code{unclass(x)}. See \code{str(x)} to examine the structure of an
object of \code{lambdas} class.
}
\seealso{
\code{\link[=parse_lambdas]{parse_lambdas()}} \code{\link[=project]{project()}}
}
\keyword{lambdas}
\keyword{maxent,}
