% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project a fitted Maxent model}
\usage{
project(lambdas, newdata, mask, quiet = FALSE)
}
\arguments{
\item{lambdas}{Either (1) a \code{MaxEnt} fitted model object (fitted with 
the \code{maxent} function in the \code{dismo} package), (2) a file path to
a Maxent .lambdas file, or (3) a \code{lambdas} object returned by 
\code{\link{parse_lambdas}}.}

\item{newdata}{A \code{RasterStack}, \code{RasterBrick}, \code{list},
\code{data.frame}, \code{data.table}, or \code{matrix} that has
layers/elements/columns whose names correspond to the names of predictors
used to fit the model. These layers/elements/columns must all have the same
length.}

\item{mask}{(Optional; requires that \code{newdata} is a \code{Raster*} 
object.) A \code{Raster} object with \code{NA} values in cells for which 
the model should \emph{not} be projected. These cells will be assigned
\code{NA} in the returned output.}

\item{quiet}{Logical. Should projection progress be reported?}
}
\value{
If \code{newdata} is a \code{RasterStack} or \code{RasterBrick}, a 
  list with three elements: 
 \itemize{
  \item{\code{prediction_raw}}{: a \code{Raster} layer giving the raw Maxent
  prediction; and}
  \item{\code{prediction_logistic}}{: a \code{Raster} layer giving the
  logistic Maxent prediction.}
  \item{\code{prediction_cloglog}}{: a \code{Raster} layer giving the
  cloglog Maxent prediction.}
}
If \code{newdata} is \emph{not} a \code{RasterStack} or \code{RasterBrick},
the raster layers will be replaced with \code{data.table}s in the returned
list.
}
\description{
Project a fitted Maxent model by predicting to new environmental data.
}
\details{
\code{project} uses feature weights described in a .lambas
  file or \code{MaxEnt} object to predict a Maxent model to environmental
  data. This function performs the projection entirely in R, without the need
  for the Maxent Java software. For tested datasets, it performs the 
  projection in roughly one third of the time taken for the same projection
  by maxent.jar.
}
\section{Warning}{

This function is still in development, and no guarantee is made for the
accuracy of its projections.
}

\examples{
# Below we use the dismo::maxent example to fit a Maxent model:
if (require(dismo) && require(rJava) && 
    file.exists(system.file('java/maxent.jar', package='dismo'))) {
  fnames <- list.files(system.file('ex', package='dismo'), '\\\\.grd$', 
                       full.names=TRUE )
  predictors <- stack(fnames)
  occurrence <- system.file('ex/bradypus.csv', package='dismo')
  occ <- read.table(occurrence, header=TRUE, sep=',')[,-1]
  me <- maxent(predictors, occ, factors='biome')

  # ... and then predict it to the full environmental grids:
  pred <- project(me, predictors)
  # This is equivalent to using the predict method for MaxEnt objects:
  pred2 <- predict(me, predictors, args='outputformat=logistic')
  pred3 <- predict(me, predictors, args='outputformat=cloglog')
  
  all.equal(values(pred$prediction_logistic), values(pred2))
  all.equal(values(pred$prediction_cloglog), values(pred3))
}
}
\references{
\itemize{
  \item{Wilson, P. W. (2009) \href{http://gsp.humboldt.edu/OLM/GSP_570/Learning Modules/10 BlueSpray_Maxent_Uncertinaty/MaxEnt lambda files.pdf}{\emph{Guidelines for computing MaxEnt model output values from a lambdas file}}.}
  \item{\emph{Maxent software for species habitat modeling, version 3.3.3k} help file (software freely available \href{https://www.cs.princeton.edu/~schapire/maxent/}{here}).}
}
}
\seealso{
\code{\link{read_mxe}}
}
\keyword{maxent,}
\keyword{predict,}
\keyword{project}
