\name{rmaxent-package}
\alias{rmaxent-package}
\alias{rmaxent}
\docType{package}
\title{
rmaxent, an R bridge to the maxent species distribution model
}

\description{
rmaxent is a R bridge to the maxent species distribution model by Steven Phillips and colleagues.
}

\details{
\tabular{ll}{
Package: \tab rmaxent\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2009-12-22\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Steven Phillips and Robert J. Hijmans
}

\keyword{ package }
\keyword{ spatial }

