% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculate Multivariate Environmental Similarity}
\usage{
similarity(x, ref, full = FALSE)
}
\arguments{
\item{x}{a \verb{Raster*}, \code{list}, \code{matrix}, or \code{data.frame}
where each layer/column/element represents focal values of an environmental
variable.}

\item{ref}{a \code{list}, \code{matrix}, or \code{data.frame} where each
column/element represents reference values for an environmental variable
(corresponding to those given in \code{x}).}

\item{full}{(logical) should similarity values be returned for all variables?
If \code{FALSE} (the default), then only the minimum similarity scores
across variables will be returned.}
}
\value{
If \code{x} is a \verb{Raster*} object, this function returns a
\code{Raster} layer giving the corresponding multivariate environmental
similarity grid. If \code{full} is \code{TRUE}, this object will be
returned in a list along with a \verb{Raster*} object giving the
environmental similarity for each layer of \code{x}. If \code{x} is a
\code{list}, \code{matrix}, or \code{data.frame}, the function will return
a vector giving multivariate similarity, or, if \code{full} is \code{TRUE},
a \code{list} with a \code{matrix} giving the environmental similarity for
each element/column of \code{x}, as well as the multivariate similarity
vector.
}
\description{
Calculate Multivariate Environmental Similarity with respect to a reference
dataset, for a set of environmental variables.
}
\details{
\code{similarity} uses the MESS algorithm described in Appendix S3
of Elith et al. 2010.
}
\examples{
library(dismo)
library(raster)
ff <- list.files(system.file('ex', package='dismo'), '\\\\.grd$', 
                 full.names=TRUE )
predictors <- stack(grep('biome', ff, value=TRUE, invert=TRUE))
occ <- read.csv(system.file('ex/bradypus.csv', package='dismo'))[, -1]
ref <- extract(predictors, occ)
mess <- similarity(predictors, ref, full=TRUE)
}
\references{
\itemize{
\item Elith, J., Kearney, M., and Phillips, S. (2010) \href{https://doi.org/10.1111/j.2041-210X.2010.00036.x}{The art of modelling range-shifting species}. \emph{Methods in Ecology and Evolution}, 1: 330-342. doi:10.1111/j.2041-210X.2010.00036.x
}
}
\keyword{environment}
\keyword{maxent,}
\keyword{mess,}
\keyword{similarity,}
