% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Identify an uncorrelated, useful subset of Maxent predictors}
\usage{
simplify(occ, bg, path, species_column = "species", response_curves = FALSE,
  logistic_format = TRUE, type = "PI", cor_thr, pct_thr, quiet = TRUE)
}
\arguments{
\item{occ}{A \code{data.frame} with predictor values for presence localities,
where columns are predictors, and rows are samples. Additionally, a column
indicating the species being modelled, which should have column name as
specified by argument \code{species_column} (can be a character string, 
numeric ID, etc.). The set of values given in this species indicator column
must be identical to the set given in the corresponding column of
\code{bg}.}

\item{bg}{A \code{data.frame} with predictor values for background
localities, where columns are predictors, and rows are samples.
Additionally, a column indicating the species being modelled, which should
have column name as specified by argument \code{species_column} (can be a
character string, numeric ID, etc.). The set of values given in this
species indicator column must be identical to the set given in the
corresponding column of \code{occ}.}

\item{path}{The output path within which output subdirectories will be 
created for each species given in the column of \code{occ} (and \code{bg}) 
specified by \code{species_column}. If missing, a temporary directory will
be used.}

\item{species_column}{The column of \code{occ} (and \code{bg}) that contains 
values indicating which species the samples belong to (e.g., this might be
species name, or species ID).}

\item{response_curves}{Logical value indicating whether response curves
should be included in Maxent model html output.}

\item{logistic_format}{Logical value indicating whether maxentResults.csv 
should report logistic value thresholds (\code{TRUE}) or cloglog value
thresholds (\code{FALSE}). This has no effect for versions of Maxent prior 
to 3.4.}

\item{type}{The variable contribution metric to use when dropping variables.
This can be \code{'PC'} (percent contribution) or \code{'PI'} (permutation
importance; the default). See the Maxent tutorial for additional details.}

\item{cor_thr}{The maximum allowable pairwise correlation between predictor
variables (calculated across presence and background localities).}

\item{pct_thr}{The minimum allowable percent variable contribution (where 
contribution type is specified by \code{type}). This should be specified as
a value between 0 and 100.}

\item{quiet}{Logical value indicating whether progress messages should be 
suppressed (\code{TRUE}) or printed (\code{FALSE}).}
}
\value{
The final fitted \code{MaxEnt} object.
}
\description{
Given a candidate set of predictor variables, this function identifies a
subset that meets specified multicollinearity criteria. Subsequently,
backward stepwise variable selection is used to iteratively drop the variable
that contributes least to the model, until the contribution of all variables 
meets a specified minimum.
}
\details{
If \code{path} is provided, subdirectories will be created within 
  \code{path}, with names equal to the values provided in the 
  \code{species_column} column of \code{occ}. Within these species
  subdirectories, two additional directories will be created: "full" contains
  the Maxent output corresponding to the model using the full uncorrelated
  subset of variables, while "final" contains the Maxent output corresponding
  to the model fit with the subset of those variables that each contribute at
  least \code{pct_thr}% to the model. Additionally, the \code{MaxEnt} R
  objects for the full and final fitted models are saved into these 
  directories, each with the name "model.rds". These can be read back into R
  with \code{\link{readRDS}}.
}
\examples{
# Below we modify the example given at ?dismo::maxent:
if (require(dismo) && require(rJava) &&
    file.exists(system.file('java/maxent.jar', package='dismo'))) {
  fnames <- list.files(system.file('ex', package="dismo"), '\\\\.grd$', 
                       full.names=TRUE)
  fnames <- grep('biome', fnames, value=TRUE, invert=TRUE)
  predictors <- scale(stack(fnames))
  occurrence <- system.file('ex/bradypus.csv', package='dismo')
  occ <- read.table(occurrence, header=TRUE, sep=',')[,-1]
  bg <- xyFromCell(predictors, Which(!is.na(sum(predictors)), cells=TRUE))
  occ_swd <- data.frame(species='bradypus', extract(predictors, occ))
  bg_swd <- data.frame(species='bradypus', extract(predictors, bg))
  m <- simplify(occ_swd, bg_swd, cor_thr=0.7, pct_thr=5, quiet=FALSE)
}
}
\keyword{correlation}
\keyword{maxent,}
\keyword{selection,}
\keyword{variable}
