% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cloglog.R
\name{to_cloglog}
\alias{to_cloglog}
\title{Transform raw or logistic values to cloglog values}
\usage{
to_cloglog(x, from, H)
}
\arguments{
\item{x}{The values to transform. This can be, e.g., a numeric atomic vector,
numeric matrix or \code{Raster*} object.}

\item{from}{A character string indicating the type of the source values. This
can be one of 'raw', 'cloglog', or 'logistic'.}

\item{H}{The entropy of the Maxent distribution. This is only used when
\code{from='raw'}.}
}
\description{
Transform Maxent raw or logistic values to cloglog (complementary log-log)
values.
}
\seealso{
\code{\link{to_logistic}}
}
\keyword{lambdas}
\keyword{maxent,}

