% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_logistic.R
\name{to_logistic}
\alias{to_logistic}
\title{Transform raw or cloglog values to logistic values}
\usage{
to_logistic(x, from, H)
}
\arguments{
\item{x}{The values to Transform. This can be, e.g., a numeric atomic vector,
numeric matrix or \code{Raster*} object.}

\item{from}{A character string indicating the type of the source values. This
can be one of 'raw', 'cloglog', or 'logistic'.}

\item{H}{The entropy of the Maxent distribution. This is only used when
\code{from='raw'}.}
}
\description{
Transform Maxent raw or cloglog (complementary log-log) values to logistic
values.
}
\seealso{
\code{\link{to_cloglog}}
}
\keyword{lambdas}
\keyword{maxent,}

