% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_maxent.R
\name{import_maxent}
\alias{import_maxent}
\title{Import a Maxent model as a MaxEnt object}
\usage{
import_maxent(dir, lambdas, html)
}
\arguments{
\item{dir}{The file path to the directory containing Maxent output.}

\item{lambdas}{The name of the .lambdas file representing the fitted model 
(e.g. \code{'species.lambdas'}), excluding the containing path. If not
provided, the first (alphabetically) .lambdas file found in \code{dir} will
be used.}

\item{html}{The name of the .html file containing reported Maxent results 
(e.g. \code{'maxent.html'}), excluding the containing path. If not 
provided, the first (alphabetically) .html file found in \code{dir} will be
used.}
}
\value{
A \code{MaxEnt} object (see \code{\link{MaxEnt-class}}).
}
\description{
Create a MaxEnt object from a previously-fitted Maxent model
}
\examples{
library(dismo)
predictors <- stack(list.files(
  file.path(system.file(package='dismo'), 'ex'), '\\\\.grd$', full=TRUE))
occ <- read.csv(file.path(system.file(package="dismo"), 'ex/bradypus.csv'))[, -1]
d <- file.path(tempdir(), 'demo')
dir.create(d)
m1 <- maxent(predictors, occ, factors='biome', path=d)
m2 <- import_maxent(d)
}

