% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting.R
\name{limiting}
\alias{limiting}
\title{Calculate which factors limit predicted habitat suitability}
\usage{
limiting(x, me)
}
\arguments{
\item{x}{A \code{RasterStack} containing predictors grids for predictors used
in the Maxent model specified at \code{me}.}

\item{me}{A a \code{MaxEnt} fitted model object (from \code{dismo::maxent}).}
}
\value{
A \code{Raster} object indicating the most limiting factor at each
  grid cell. This is defined, for each grid cell, as the variable whose value
  is most responsible for decreasing suitability at that cell. The decrease 
  in suitability is calculated, for each predictor in turn, relative to the 
  suitability (logistic prediction) that would be achieved if that predictor
  took the value equal to the mean at occurrence sites (median for
  categorical variables). The predictor associated with the largest decrease
  in suitability is the most limiting factor.
}
\description{
Calculate a limiting factor surface corresponding to projections of a Maxent
model.
}
\examples{
library(dismo)
library(rJava)

fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''),
                     pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
occurence <- paste(system.file(package='dismo'), '/ex/bradypus.csv', sep='')
occ <- read.table(occurence, header=TRUE, sep=',')[,-1]
me <- maxent(predictors, occ, factors='biome')

limiting(predictors, me)
}
\references{
Elith, J., Kearney, M. and Phillips, S. (2010), \href{http://onlinelibrary.wiley.com/doi/10.1111/j.2041-210X.2010.00036.x/abstract}{The art of modelling range-shifting species.} \emph{Methods in Ecology and Evolution}, 1: 330–342. doi: 10.1111/j.2041-210X.2010.00036.x
}

