% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaxent.R
\docType{package}
\name{rmaxent-package}
\alias{rmaxent}
\alias{rmaxent-package}
\title{rmaxent: Tools for working with Maxent in R}
\description{
An R implementation of tasks related to 'Maxent' species distribution modelling: projecting models; computing information criteria; creating limiting factors maps; importing mxe files; recreating MaxEnt class objects and parsing .lambdas files.
}
\author{
\strong{Maintainer}: John B. Baumgartner \email{johnbaums@gmail.com}

Authors:
\itemize{
  \item Peter D. Wilson
}

}
\keyword{internal}
