% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculate Multivariate Environmental Similarity}
\usage{
similarity(x, ref, full = FALSE)
}
\arguments{
\item{x}{a \verb{Raster*}, \code{list}, \code{matrix}, or \code{data.frame}
where each layer/column/element represents focal values of an environmental
variable.}

\item{ref}{a \code{list}, \code{matrix}, or \code{data.frame} where each
column/element represents reference values for an environmental variable
(corresponding to those given in \code{x}).}

\item{full}{(logical) should similarity values be returned for all variables?
If \code{FALSE} (the default), then only the minimum similarity scores
across variables will be returned.}
}
\value{
If \code{x} is a \verb{Raster*} object, this function returns a list
containing:
\itemize{
\item \code{similarity}: a \code{RasterStack} giving the environmental similarities for
each variable in \code{x} (only included when \code{full=TRUE});
\item \code{similarity_min}: a \code{Raster} layer giving the minimum similarity value
across all variables for each location (i.e. the MESS);
\item \code{mod}: a factor \code{Raster} layer indicating which variable was most
dissimilar to its reference range (i.e. the MoD map, Elith et al. 2010);
and
\item \code{mos}: a factor \code{Raster} layer indicating which variable was most
similar to its reference range.
}

If \code{x} is a \code{list}, \code{matrix}, or \code{data.frame}, the function will return
a list as above, but with \code{RasterStack} and \code{Raster} objects replaced by
matrix and vectors.
}
\description{
Calculate Multivariate Environmental Similarity and most dissimilar/similar
variables with respect to a reference dataset, for a set of environmental
variables.
}
\details{
\code{similarity} uses the MESS algorithm described in Appendix S3
of Elith et al. 2010.
}
\examples{
library(dismo)
library(raster)
ff <- list.files(system.file('ex', package='dismo'), '\\\\.grd$', 
                 full.names=TRUE )
predictors <- stack(grep('biome', ff, value=TRUE, invert=TRUE))
occ <- read.csv(system.file('ex/bradypus.csv', package='dismo'))[, -1]
ref <- extract(predictors, occ)
mess <- similarity(predictors, ref, full=TRUE)

\dontrun{
library(rasterVis)
library(RColorBrewer)
levelplot(mess$mod, col.regions=brewer.pal(8, 'Set1'))
levelplot(mess$mos, col.regions=brewer.pal(8, 'Set1'))
}
}
\references{
\itemize{
\item Elith, J., Kearney, M., and Phillips, S. (2010) \href{https://doi.org/10.1111/j.2041-210X.2010.00036.x}{The art of modelling range-shifting species}. \emph{Methods in Ecology and Evolution}, 1: 330-342. doi:10.1111/j.2041-210X.2010.00036.x
}
}
\keyword{environment}
\keyword{maxent,}
\keyword{mess,}
\keyword{similarity,}
